+++ CliReader/PolyLine.h
@@ -13,7 +13,7 @@ public:
 
 	unsigned int m_id;
 	unsigned int m_dir;
-	unsigned int m_nPoints;
+	size_t m_nPoints;
 	std::vector<double> m_points;
 	double polygonArea2D() const;
 	void polygonBBox(std::array<double, 4>& BBox2d) const;
